[DataLayer] Prepare beta version of the Adobe Client Data Layer
- move the validation code of the item to the Item class
